LayoutTests/imported/w3c/web-platform-tests/eventsource/request-cache-control-expected.txt

Blocked access to external URL http://www2.localhost:8800/eventsource/resources/cors.py?run=cache-control
CONSOLE MESSAGE: EventSource cannot load http://www2.localhost:8800/eventsource/resources/cors.py?run=cache-control due to access control checks.

Harness Error (TIMEOUT), message = null

PASS EventSource: Cache-Control
TIMEOUT EventSource: Cache-Control 1 Test timed out
PASS EventSource: Cache-Control 2
